fix: segfault

+++ VleerhondApp/src/midi/midi_channel.cpp
@@ -42,7 +42,7 @@ namespace Vleerhond
     void MidiChannel::processNoteEvents()
     {
         uint32_t time = Utils::millis();
-        for (size_t i = this->nr_of_events - 1; i >= 0; i--)
+        for (int16_t i = this->nr_of_events - 1; i >= 0; i--)
         {
             NoteEvent event_i = this->events[i];
             if (event_i.time <= time)
@@ -141,7 +141,7 @@ namespace Vleerhond
                 this->data[i].length -= 1;
             }
         }
-        for (size_t i = (int)this->size - 1; i >= 0; i--)
+        for (int16_t i = (int)this->size - 1; i >= 0; i--)
         {
             if (this->data[i].length == 0)
             {
@@ -155,7 +155,7 @@ namespace Vleerhond
 
     void MidiChannel::untieNotes()
     {
-        for (size_t i = (int)this->size - 1; i >= 0; i--)
+        for (int16_t i = (int)this->size - 1; i >= 0; i--)
         {
             if (!this->pedal)
             {
@@ -169,7 +169,7 @@ namespace Vleerhond
 
     void MidiChannel::allNotesOff()
     {
-        for (size_t i = (int)this->size - 1; i >= 0; i--)
+        for (int16_t i = (int)this->size - 1; i >= 0; i--)
         {
             noteOff(this->data[i].pitch);
         }
